pyproject.toml
[project]
name = "tcp-demo"
version = "0.1.0"
description = "TCP/IP Transport Layer Demonstration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
]

[tool.uv]
dev-dependencies = [
    "mypy>=1.13.0",
    "pyright>=1.1.386",
    "ruff>=0.7.0",
]
